backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
asyncpg==0.29.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Security - Authentication & JWT
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6

# Security - Encryption (Fineract-compliant AES-256)
cryptography==41.0.7

# Security - 2FA/TOTP
pyotp==2.9.0
qrcode==7.4.2

# Rate Limiting & Caching
redis==5.0.1

# Task Queue & Scheduler
#celery==5.3.4
# flower==2.0.1  # Celery monitoring

# Scheduler - APScheduler for recurring jobs
apscheduler==3.10.4

# Testing
pytest==7.4.3
httpx==0.25.2

# Utilities
python-dateutil==2.8.2
aiofiles==23.2.1
websockets==12.0

# Reporting
reportlab==4.0.7
openpyxl==3.1.2
pandas==2.1.3
numpy==1.26.2
fpdf2==2.7.9

# Security - Input Validation
email-validator==2.1.0
phonenumbers==8.13.26

# Security - Password Strength
zxcvbn==4.4.28

# HTTP Client for Webhooks
aiohttp==3.9.1

# Date utilities
python-dateutil==2.8.2